refactor: Mapper 패키지 분리 및 관련 코드 정리
